// Cyberpunk Red NPC Sheet/src/types/skills.types.ts
import type CharacterSkillStructure from '@/classes/characterSkills.classes'

export type AwarenessSkillsType = {
    Concentration: CharacterSkillStructure;
    ConcealRevealObject: CharacterSkillStructure;
    LipReading: CharacterSkillStructure;
    Perception: CharacterSkillStructure;
    Tracking: CharacterSkillStructure;
};

export type BodySkillsType = {
    Athletics: CharacterSkillStructure
    Contortionist: CharacterSkillStructure
    Dance: CharacterSkillStructure
    Endurance: CharacterSkillStructure
    ResistTortureDrugs: CharacterSkillStructure
    Stealth: CharacterSkillStructure
}

export type ControlSkillsType = {
    DriveLandVehicle: CharacterSkillStructure
    PilotAirVehicle: CharacterSkillStructure
    PilotSeaVehicle: CharacterSkillStructure
    Riding: CharacterSkillStructure
}

export type EducationSkillsType = {
    Accounting: CharacterSkillStructure
    AnimalHandling: CharacterSkillStructure
    Bureaucracy: CharacterSkillStructure
    Business: CharacterSkillStructure
    Composition: CharacterSkillStructure
    Criminology: CharacterSkillStructure
    Cryptography: CharacterSkillStructure
    Deduction: CharacterSkillStructure
    Education: CharacterSkillStructure
    Gamble: CharacterSkillStructure
    LibrarySearch: CharacterSkillStructure
    Tactics: CharacterSkillStructure
    WildernessSurvival: CharacterSkillStructure
}

export type FightingSkillsType = {
    Brawling: CharacterSkillStructure
    Evasion: CharacterSkillStructure
    MartialArts: CharacterSkillStructure
    MeleeWeapon: CharacterSkillStructure
}

export type PerformanceSkillsType = {
    Acting: CharacterSkillStructure
}

export type RangedWeaponSkillsType = {
    Archery: CharacterSkillStructure
    AutoFire: CharacterSkillStructure
    Handgun: CharacterSkillStructure
    HeavyWeapons: CharacterSkillStructure
    ShoulderArms: CharacterSkillStructure
}

export type SocialSkillsType = {
    Bribery: CharacterSkillStructure
    Conversation: CharacterSkillStructure
    HumanPerception: CharacterSkillStructure
    Interrogation: CharacterSkillStructure
    Persuasion: CharacterSkillStructure
    PersonalGrooming: CharacterSkillStructure
    Streetwise: CharacterSkillStructure
    Trading: CharacterSkillStructure
    WardrobeStyle: CharacterSkillStructure
}

export type TechniqueSkillsType = {
    AirVehicleTech: CharacterSkillStructure
    BasicTech: CharacterSkillStructure
    Cybertech: CharacterSkillStructure
    Demolitions: CharacterSkillStructure
    ElectronicsSecurityTech: CharacterSkillStructure
    FirstAid: CharacterSkillStructure
    Forgery: CharacterSkillStructure
    LandVehicleTech: CharacterSkillStructure
    PaintDrawSculpt: CharacterSkillStructure
    Paramedic: CharacterSkillStructure
    PhotographyFilm: CharacterSkillStructure
    PickLock: CharacterSkillStructure
    PickPocket: CharacterSkillStructure
    SeaVehicleTech: CharacterSkillStructure
    WeaponsTech: CharacterSkillStructure
}